backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic>=2.8.2,<2.9.0
python-multipart==0.0.9
langchain==0.1.11
langchain-openai==0.0.8
langchain-google-genai==0.0.8
pypdf==4.1.0
mistralai>=1.0.1
requests==2.31.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator
bcrypt==4.0.1
celery
redis
